+++ containers/VideoDetail/hooks/index.ts
@@ -1,2 +1,3 @@
 export { default as useFetchCommentById } from './useFetchCommentById'
 export { default as useFetchDiscoverById } from './useFetchDiscoverById'
+export { default as usePostSocketComment } from './usePostSocketComment'
